add CommandStartup support for non Web ASP.NET Core apps,
