+++ MariCommands/Extensions/MariCommandsDependencyInjectionExtensions.cs
@@ -1,8 +1,13 @@
 using System;
 using System.Linq;
+using MariCommands.Builder;
+using MariCommands.Factories;
+using MariCommands.Parsers;
+using MariCommands.Providers;
 using MariCommands.TypeParsers;
 using MariGlobals.Extensions;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 
 namespace MariCommands.Extensions
@@ -55,7 +60,7 @@ namespace MariCommands.Extensions
         /// <typeparam name="TParser">The type parser to be injected.</typeparam>
         /// <typeparam name="T">The type of this type parser can parse.</typeparam>
         /// <returns>The current service collection.</returns>
-        public static IServiceCollection AddTypeParser<TParser, T>(this IServiceCollection services, bool createNullable)
+        public static IServiceCollection AddTypeParser<TParser, T>(this IServiceCollection services, bool createNullable = true)
             where T : struct
             where TParser : class, ITypeParser<T>
         {
@@ -93,7 +98,7 @@ namespace MariCommands.Extensions
         /// <typeparam name="TParser">The type parser to be injected.</typeparam>
         /// <typeparam name="T">The type of this type parser can parse.</typeparam>
         /// <returns>The current service collection.</returns>
-        public static IServiceCollection TryAddTypeParser<TParser, T>(this IServiceCollection services, bool createNullable)
+        public static IServiceCollection TryAddTypeParser<TParser, T>(this IServiceCollection services, bool createNullable = true)
             where T : struct
             where TParser : class, ITypeParser<T>
         {
@@ -110,7 +115,7 @@ namespace MariCommands.Extensions
         /// <param name="createNullables">If this lib will create nullable type parsers for these 
         /// primitive type parsers.</param>
         /// <returns>The current service collection.</returns>
-        public static IServiceCollection AddPrimitiveTypeParsers(this IServiceCollection services, bool createNullables)
+        public static IServiceCollection AddPrimitiveTypeParsers(this IServiceCollection services, bool createNullables = true)
         {
             services.TryAddTypeParser<PrimitiveTypeParser<char>, char>(createNullables);
             services.TryAddTypeParser<PrimitiveTypeParser<bool>, bool>(createNullables);
@@ -125,6 +130,66 @@ namespace MariCommands.Extensions
             services.TryAddTypeParser<PrimitiveTypeParser<float>, float>(createNullables);
             services.TryAddTypeParser<PrimitiveTypeParser<double>, double>(createNullables);
             services.TryAddTypeParser<PrimitiveTypeParser<decimal>, decimal>(createNullables);
+
+            return services;
+        }
+
+        /// <summary>
+        /// Add the default string type parser to the dependency.
+        /// </summary>
+        /// <param name="services">The current service collection.</param>
+        /// <returns>The current service collection.</returns>
+        public static IServiceCollection AddStringTypeParser(this IServiceCollection services)
+            => services.TryAddTypeParser<StringTypeParser, string>();
+
+        /// <summary>
+        /// Add the default enum type parser to the dependency.
+        /// </summary>
+        /// <param name="services">The current service collection.</param>
+        /// <returns>The current service collection.</returns>
+        public static IServiceCollection AddEnumTypeParser(this IServiceCollection services)
+            => services.TryAddTypeParser<EnumTypeParser, Enum>();
+
+        /// <summary>
+        /// Add all default type parsers to the dependency.
+        /// </summary>
+        /// <param name="services">The current service collection.</param>
+        /// <param name="createNullables">If this lib will create nullables type parsers for these type parsers.</param>
+        /// <returns>The current service collection.</returns>
+        public static IServiceCollection AddAllDefaultTypeParsers(this IServiceCollection services, bool createNullables = true)
+        {
+            services.AddPrimitiveTypeParsers(createNullables);
+            services.AddStringTypeParser();
+            services.AddEnumTypeParser();
+
+            return services;
+        }
+
+        /// <summary>
+        /// Adds all basic services of the MariCommands for the dependency.
+        /// </summary>
+        /// <param name="services">The current service collection.</param>
+        /// <param name="addAllDefaultTypeParsers">If this lib wil inject all default type parsers to the dependency.</param>
+        /// <param name="createNullables">If this lib will create nullables type parsers for these type parsers.</param>
+        /// <returns>The current service collection.</returns>
+        public static IServiceCollection AddBasicMariCommandsServices(this IServiceCollection services, bool addAllDefaultTypeParsers = true, bool createNullables = true)
+        {
+            services.TryAddSingleton<ICommandServiceOptions, CommandServiceOptions>();
+            services.TryAddSingleton<IContextExecutor, ContextExecutor>();
+            services.TryAddSingleton<IModuleCache, ModuleCache>();
+
+            services.TryAddTransient<ICommandService, CommandService>();
+            services.TryAddTransient<ICommandApplicationBuilderFactory, CommandApplicationBuilderFactory>();
+            services.TryAddTransient<IModuleConfigurer, ModuleConfigurer>();
+            services.TryAddTransient<IModuleFactory, ModuleFactory>();
+            services.TryAddTransient<ICommandFactory, CommandFactory>();
+            services.TryAddTransient<IParameterFactory, ParameterFactory>();
+            services.TryAddTransient<IArgumentParser, ArgumentParser>();
+            services.TryAddTransient<ITypeParserProvider, TypeParserProvider>();
+            services.TryAddTransient<ICommandExecutorProvider, CommandExecutorProvider>();
+
+            if (addAllDefaultTypeParsers)
+                services.AddAllDefaultTypeParsers();
 
             return services;
         }
